Refactor 편지 작성, 편지 임시저장 분리
